services/validation-worker/requirements.txt
nats-py
psycopg[binary] 
boto3 
boto3-stubs
boto3-stubs[essential]
trimesh 
numpy 
Pillow 
pydantic 
pydantic_settings
types-aioboto3
puremagic
pytest
asyncpg
pytest-asyncio